requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
black==25.1.0
bleach==6.2.0
click==8.1.8
colorama==0.4.6
cssmin==0.2.0
Django==5.2.5
django-appconf==1.1.0
django-compressor==4.5.1
django-jazzmin==3.0.1
django-maintenance-mode==0.22.0
django-multi-captcha-admin==2.0.0
django-ranged-response==0.2.0
django-rest-passwordreset==1.5.0
django-robots==6.1
django-simple-captcha==0.6.2
django-summernote==0.8.20.0
django-taggit==6.1.0
GoogleAppEngineCloudStorageClient==1.9.22.1
gunicorn==23.0.0
jsmin==3.0.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
python-decouple==3.8
python-fsutil==0.15.0
rcssmin==1.1.2
rjsmin==1.2.2
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.1
webencodings==0.5.1
